Starting on bank and chips.

+++ Backend/Cycino/src/main/java/coms309/Cycino/Games/Blackjack/Blackjackcontroller.java
@@ -71,7 +71,7 @@ public class Blackjackcontroller {
 
     @PutMapping("/blackjack/bet/{lobby}/{userId}/{bet}")
     public Map<String, Object> addBet(@PathVariable long lobby, @PathVariable long userId, @PathVariable double bet){
-        return bjs.setBet(lobby, userId, bet);
+        return bjs.setBet(lobby, userId, (int) bet);
     }
 
 }
